build FileScubaLogger from a config file
Summary:
Workers run out-of-proc, and we want them to have loggers with the same configuration as our orchestrator. To do so, we want to be able to create loggers from a common config file on disk.

For now, we have 3 logger modes:
1. FileScuba (the intended production mode)
2. FileOnly (for local testing purposes)
3. Dummy (for performance testing without logging)

I've kept the guards and logger separate to resemble `slog::Logger`'s design. There might be value in combining the two -- though when trying it, I noticed that there's a lot of additional work required to add lifetimes and change function calls everywhere. We'd have to rework at least a few APIs, including `hh_decl_shmem::connect`. I didn't think this work was worth it, so I've abandoned that route for now.

Reviewed By: ljw1004

Differential Revision: D42870896

fbshipit-source-id: 26d65bc1a2fd80f49393d709aa93154915c1d7f4