Mark typename and classname NoAutoDynamic in hhi
Summary: We don't want to restrict the use of `classname` and `typename` to classes that support dynamic. By marking the definitions as `<<__NoAutoDynamic>>` the pessimiser (explicit and implicit) won't add a `supportdyn<mixed>` bound to the generic parameter. Note that because `typename<_>` has `string` as its upper bound, both `classname` and `typename` automatically support dynamic type, irrespective of their generic parameter.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D42926852

fbshipit-source-id: b7961e6003eda7493817cb63ed45fff812bd4d17