Remove redundant RequireDynamic]
Summary: For generic types under Sound Dynamic we have a range of options. One is the use of `__RequireDynamic` on generic parameters, which means that the generic type is dynamic-aware subtype of `dynamic` only if the parameter is *exactly* `dynamic`. This is not appropriate for a uniform pessimisation policy, because `Vector<string>` and similar then do not support dynamic. It was a mistake to put these on generics in `hhi` - let's remove them.

Reviewed By: periodic1236

Differential Revision: D43003832

fbshipit-source-id: abe0c321a821c56a15df36e6c1b99a2160186547