Reflect Provider_context entries in hackrs_provider_backend
Summary:
In our implementations of FileProvider and NamingProvider, access IDE buffers in the Provider_context.t to return results reflecting their contents.

This is awkward because Provider_context depends on Rust_provider_backend, so we cannot use the type Provider_context.t in Rust_provider_backend. Instead, when the context is non-empty, accept a value of type ctx_proxy containing closures which capture the Provider_context.t.

Reviewed By: shayne-fletcher

Differential Revision: D42945761

fbshipit-source-id: 9fa85c804115211907599c7643eb228e673f2ef7