Remove the ref `argument_global_type` from autocompleteService
Summary:
Previously, we relied on this ref being set to decide whether we should do more completion.

This isn't sufficient: there are syntactic positions where we could complete to either a keyword or a function call, so we should keep going.

This logic was also extremely error prone. We relied on `Acprop` to decide if we should do local completion. This was confusing (apparently `Acprop` does not relate to properties) and prone to bugs (e.g. see D43014507 (https://github.com/facebook/hhvm/commit/18aff16dd03429b769c9f689a164611e61c597e6)).

Instead, compute locals completion as soon as we find the cursor position, and remove this ref entirely.

Reviewed By: vsiles

Differential Revision: D43015840

fbshipit-source-id: b25b0da94b05d30ed993ffe0b7eb43ef22b7b01d