make getSymbolValue and getAddressValue return pointers
Summary:
Strobelight sees a lot of crashes due to attempting to get the value of Elf symbols that have been moved/removed by Elf manipulation. We've tried to go after the source of these issues but still see this issue a lot: https://pxl.cl/2rL38

We can change the API of `getSymbolValue` and `getAddressValue` as this is primarily used by Strobelight and is within the experimental folder. These changes remove the FOLLY_SAFE_CHECKs so Strobelight doesn't exit if we fail to get the value for an address or a symbol.

Reviewed By: nslingerland, luciang

Differential Revision: D42990975

fbshipit-source-id: 08294ed7e9e6cb671a370ff8bda426d674b9625b