add a command to get the log path
Summary:
Sometimes the logs in the diag are not enough and I have to tell users to go
get me more of their logs.

This is much easier todo if we can just point them directly to the log path.

I also locally want to read the file sometimes and it can be annoying to find
because I forget the log path.

Adding a simple cli to print the log path. we could do better by adding
uploading capabilities and all like the `eden debug log` command, but that
will have to be a later diff.

Reviewed By: fanzeyi

Differential Revision: D43070266

fbshipit-source-id: 47b50f502dad36c8d823c45665ffdcda5757ae64