Port `Naming_elab_class_id`
Summary:
Reimplementation of `Naming_elab_class_id`; the pass expects `ClassId`s of the for `(_,_,CIexpr(_,_,Id | Lvar))` from lowering.

Where the Id is a name refering to a class (i.e. `self`, `parent` or `static`) we replace with the canonical representation when inside a class or subsitute for the unknown class name if not and raise an error.

Where we have an Lvar, we elaborate `Lvar(_,this)` into `This`. This overlaps with lvar elab and should be revisited.

At the moment, this pass does not enforce the invariant about which `ClassId` ctors are expected from lowering and will silently continue.

Reviewed By: shayne-fletcher

Differential Revision: D43082048

fbshipit-source-id: 6f6a1a023422069731efb264b1a9a64df474b9cc