Thread TAST variation under dynamic assumptions
Summary:
When a function or a class supports dynamic type, the function and the methods in the class are checked under dynamic assumptions (that is the parameters and the return are given the `dynamic` type and the body is checked in that environment). This effectively leads to alternative TASTs for functions and methods. So far we have been doing the check and ignoring the resulting TAST variations.

This diff adds those variations back into TASTs. So if a function `f` is in the source code and it we are running in SDT mode with `--tast-under-dynamic`, the TAST would contain two definitions for `f` one checked under normal assumptions and another under normal assumptions. For a method the classes contains two method TASTs for the method (rather than two classes of the same name).

Reviewed By: andrewjkennedy

Differential Revision: D42806545

fbshipit-source-id: d3e6ece5d8b39b5f8ae4c44bc744dfc6cc65c9b8