FileScubaLogger plumbing to hh_decl
Summary:
This diff adds FFI methods to create a `FileScubaLogger` from a config file in OCaml land. I've plumbed through to `connect` in `hh_decl_ffi` as a start.

Next steps are to plumb the logger through to `hh_fanout` and to expose FFI methods for the OCaml worker to use the logger.

Reviewed By: ljw1004

Differential Revision: D42972545

fbshipit-source-id: 0f34f7355cf60ea19c2e97493f9b1e849ccb8ebc