Rename common::by_ref_context to common::context, gen_visitor::Args to common::args::Args
Summary:
The simplified context used in gen_by_ref_decl_visitor is good enough for other use cases like generating the elaboration transform, so rename it from by_ref_context to context.

The arguments used for gen_visitor are the same as those that will be used for generating the elaboration transform, so rename gen_visitor::Args to common::args::Args.

Reviewed By: shayne-fletcher

Differential Revision: D43097013

fbshipit-source-id: 66b08c8179233572ce9e8e8946d3a76c424cbe32