Pass async lambda to expected type dynamic
Summary: When passing a lambda where the type `dynamic` is expected, we assume parameters and return on the lambda to have the type `dynamic`. Because subtyping with non-explicit return types uses ordinary subtyping, not dynamic-aware subtyping, this means that we get errors (e.g. we can't type `return 1;`) that we wouldn't get with an explicit return type. Solution is to actually to infer return type in the case that it's missing: it's only the parameter types that we need here.

Reviewed By: vassilmladenov

Differential Revision: D43082008

fbshipit-source-id: ab800ddf3cd90cc2f44bb5751921ccb6001c2d0b