Don't negate int64 min
Summary:
This is undefined behavior, and Clang with LTO leverages the undefined behavior
to simplify the code in a way that breaks tests. So special case the handling of
int64 min in UFixedToFloat and UFixedToDouble and remove the attribute to
prevent inlining of CountLeadingZeros that was being used to fix the test by
preventing Clang w/ LTO from optimizing the code.

Reviewed By: edwinsmith

Differential Revision: D43165435

fbshipit-source-id: aef721733da84b539241e8612bf799f681c942c6