Add `Default` bound on `Ex` type param
Summary:
During some elaboration phases we end up generating expressions, typically to mark semantically invalid or missing expressions.

In order to do this, we must be able to generate an `Ex` annotation for this expression.

This diff adds a `Default` trait bound to this type parameter. This is appropriate since the choice of annotation must arbitrary for the elaboration to be meaningful.

In practice, the `Ex` paramteter is `()` throughout elaboration.

Reviewed By: shayne-fletcher

Differential Revision: D43115058

fbshipit-source-id: a864ba90be43acdfc618afb97c5381587ffcfb62