Port `Naming_elab_collection`
Summary:
Reimplements `Naming_elab_collection`; the lowerer represents collection literals uniformally as `Collection`. This pass determines whether we have a `ValCollection`, `KeyValCollection` or `Pair` based on the name and subsequently elaborates / validates any explicit type args and the expressions within the collection literal.

We wrap with the Invalid marker when:
- the name is not a collection kind
- we have a pair with more than or fewer than 2 elements

Reviewed By: shayne-fletcher

Differential Revision: D43118432

fbshipit-source-id: e10d0eb78c1199c10962ee81c21a98fa88347f6c