Gate reified param optimization
Summary:
The only "production" use of DDB, today, is gating an optimization around param verification. If we know a parameter doesn't contain reified generics, we can elide adding a check for such (I believe HHBBC and the JIT would do the same later on).

We want `EnableDecl` to be unobservable during the initial rollout, so let's add a flag for this specific optimization.

Reviewed By: shayne-fletcher

Differential Revision: D42197211

fbshipit-source-id: cf78f7d92301aa9fb01a11e643c9dd34dee91f35