Don't loosen types in init_return_type()
Summary:
When calculating initial function return types, we build a vec type
for in-out functions. However, we immediately then call loosen_all()
on the type, dropping the specialization! There's no need to call
loosen_all() on the return type. After this was done, it turns out
there was a bug while building the vec. We were building it in the
wrong order. Fix that.

Reviewed By: kavoor

Differential Revision: D43180403

fbshipit-source-id: aa97f3d6f7cd6ef3c7cb42b2f2240e68d7ea7064