Dump loop iteration stats
Summary:
It would be nice to be able to programmatically grep for all the loop iteration stats in tcprint.

This diff just takes the loop header, and finds all predecessors that are not dominated by the loop header.
Its not exact, but should give us a decent idea of whether these loops iterate 1-2 times on average, or more like 8-10.

This diff should make it easy to grep tcprint to get rough numbers for all translations. If it looks good, I can investigate more.

Reviewed By: ottoni

Differential Revision: D43202195

fbshipit-source-id: b9cd8516d66a3e6f7823b099e28beadc31db4d8b