Merge transform and pass crates into single elab crate
Summary:
It's very expensive to compile a use of the `Transform` trait: since we do not use `&dyn`, the traversal and passes are all instantiated at the call site. This is good for inlining, but bad for compile times and code size: many uses of the `Transform` trait means many different instantiated versions of the traversal code.

We ran into this in the FFP as well, and addressed it by deliberately choosing a few SmartConstructors parameterizations and instantiating them once in the parser/api crates. Similarly, we could avoid exposing `Transform`/`Pass`/`Passes` from elab at all, exposing only the functions that are needed outside the crate (like `elaborate_program`).

This doesn't restrict us from writing transforms with just a few passes in tests; it just constrains those tests to be unit tests in the elab crate (as they already are) so that they can access the `Transform`/`Pass`/`Passes` symbols.

Reviewed By: enetsee

Differential Revision: D43206210

fbshipit-source-id: 405aabd6f18de6f219a256ad10a54e6fb0d4c780