Store EnumClassLabel in attribute arguments
Summary:
Previously, the lowerer would transform `<<__Memoize(#Foo)>>` or `<<__Memoize(Bar#Foo)>>` to `<<__Memoize("Foo")>>`.

This prevented the rest of the toolchain seeing the enum class label, so no type checking or IDE support was offered.

Preserve the enum class label during lowering, and only perform the transformation to `"Foo"` in the emitter.

This preserves the existing parse errors, so `__Memoize(#NoSuchLabel)` is still a parse error. It also now treats `__Memoize("foo")` as a parse error.

Reviewed By: viratyosin

Differential Revision: D43058999

fbshipit-source-id: 405d5b9f745006f91657cf8f5cd89be2362f8f8f