Like type logger
Summary:
This diff adds a logger for like types.

It is triggered by setting "like_type" log level to a number greater than 0.

The definition of like type for a type `t` is that `dynamic <: t` and `mixed </: t`.

The logger looks at every expression and also certain decl constructs, i.e., properties, parameters, and returns.

Reviewed By: vassilmladenov

Differential Revision: D43161192

fbshipit-source-id: 80b15affa500bd2c1b67c65a2e345d7ea52f4644