Replace `fun` and `class_meth` with function pointers in `hh` tests
Summary:
`fun`, `class_meth`, and `inst_meth` ought to be properly ripped out of Hack. The steps here are roughly:
1. Codemod away the syntax from tests (we are here)
2. Default the options for disabling these at the parser level to `on`
3. Rip out the typechecker & compiler support
4. Relax the parser rules in (2)

This diff takes a stab at the "trivial" `class_meth` and `fun` in typechecker tests.  My general rule here was to let the codemod do its thing if the automated change looked like it was testing *roughly* the same thing.  There are some tests that look a bit weird now.

Reviewed By: andrewjkennedy

Differential Revision: D42255400

fbshipit-source-id: 5ded99ff50f358635e35fa8c88ee37dd47dcb35c