use the sub_supportdyn parameter more uniformly
Summary:
^

In preparation for enabling `supportdyn<vec<t>> <: vec<supportdyn<t>>`. In subtyping, we track when we've gone under supportdyn on the left, and if we add a `IsSubtype` constraint, we wrap the type in supportdyn (unless it's a variable, we don't want to get in the way of solving things). There are two additional heuristics: when solving for equality, we strip `supportdyn` from the lower bound so that a # variable with `supportdyn<t>` lower bound and `t` upper bound gets solved; and we reduce `supportdyn<t> <: supportdyn<u>` to `t <: u` without remembering that we're in supportdyn on the left. This seems less complete, but setting `sub_supportdyn` caused test/typecheck/anon14.php to crash building an every increasing series of type variables wrapped with supportdyn. It would be nice to fix that.

Reviewed By: andrewjkennedy

Differential Revision: D43274598

fbshipit-source-id: 60db1a7dc05628affee15c5403214ae683585ebd