Follow Rust naming conventions in validate_class_req
Summary: I think the formula for these names has been applied a little too strictly! It's true that this pass deals with a type named `Class_`, but it also deals with "classes" in the abstract sense, and I would interpret these names in the latter sense rather than the former (i.e., the name should say "this is a pass dealing with requirements on classes" rather than "this is a pass dealing with requirements on `Class_`es").

Reviewed By: shayne-fletcher

Differential Revision: D43337023

fbshipit-source-id: 4f7eea481209d5e555bf5378223eed695e6939f8