Expression Tree Blocks
Summary:
# Motivation
Expression trees are of the form:

```
Code`<expr>`
```
Where `Code` is the name of expression tree class we are creating an instance of and `<expr>` is some valid Hack expression. There are times when it is desirable to include statements as part of the expression tree. For instance, if the user wants to assign a spliced value to a variable since it will be referred to multiple times in the expression like in the following example:

```
Code`
  $name = ${<expr>};
  if (CodeString::is_empty($name)) {
    $name = default_name();
  }
  process_string($name)
`
```
Here you assign the result of splicing an expression to a variable within the expression tree `$name`. You check if it is empty using a helper function defined for the expression tree that you set to some default. Afterwards you further process the string. This is not valid syntax for an expression tree, instead the statements need to be wrapped in a lambda that is then immediately invoked, i.e.

```
Code`() ==> {
  $name = ${<expr>};
  if (CodeString::is_empty($name)) {
    $name = default_name();
  }
  return process_string($name);
}()`
```

This extra noise hurts the ergonomics of expression trees and this feature hopes to alleviate it.
# Proposal
The proposed solution is to provide sugar for the immediately invoked lambda pattern by introducing the notion of expression tree blocks. Blocks in expression trees would be of the form:
```
Code`{<statements>}`
```
This sugar is expanded during lowering to the form:
```
Code`() ==> { <statements> }()`
```
This means the example from before could be written as:
```
Code`{
  $name = ${<expr>};
  if (CodeString::is_empty($name)) {
    $name = default_name();
  }
  return process_string($name);
}`
```
# Alternatives Considered
Expression trees were inspired by languages that are largely expression based, namely OCaml and Scala. In those languages something like expression tree blocks would not be needed. We could make Hack expression based as well. This would be a much more invasive change to the language and would need a stronger motivation than expression trees.

Another option is to add a general way of utilizing statements as an expression. An example of this can be seen in [do expressions](https://github.com/tc39/proposal-do-expressions) which are being proposed for JavaScript. There have been previous discussions in the past of including a similar feature in the language as well in the form of compound expressions. This proposal does not prevent us from pursuing such features in the future if desired, and prefer to go with a feature more localized to expression trees.

# Implementation
The following changes were made to the language:
* The ffp schema for `PrefixedCodeExpression` was changed by renaming the "expression" field to "body"
* The expression parser was changed so that instead of expecting an expression after a backtick we will parse a lambda body (the same parsing we do after `==>`, i.e. `() ==> <expression>` or `() ==> { <statement> }`)
* During lowering, if the body of the PrefixedCodeExpression is a compound statement we lower it to lambda expression that is immediately invoked in the AST

Since the translation is done during lowering and before the other desugaring done on expression trees no further changes are needed, namely the type checker and emitter are not changed.

# Gating
I am gating this as an ongoing release feature. This means the type checker will require including `<<file:_EnableUnstableFeatures('expression_tree_blocks')>>`, but the runtime will allow the feature without the attribute being present. This is the minimal amount of gating required because this is a parser change. If we did not have this gating then it is possible for the type checker to allow usage of the feature before the runtime release is available.

I'll also argue the gating is sufficient (in other words that we do not need to be more restrictive). The feature is simple enough we do not gain any benefit from making it an unstable feature.

Reviewed By: hgoldstein

Differential Revision: D43251333

fbshipit-source-id: 5df5cfabbf5f7c136e8e3c16b2f652df58b25256