Implement supportdyn pushing in subtyping
Summary:
We want to remove the `as supportdyn<mixed>` bounds from all Containers/KeyedContainers that were used for implicit pessimisation.

This will weaken refinement because when we do `$c is KeyedContainer<_,_>` then the old bound would be applied to the type variables introduced by `is`. So if `$c : supportdyn<mixed>` then the `is` would refine it to `$c : supportdyn<KeyedContainer<T#1, T#2>> where T#2 as supportdyn<mixed>`. Now it will just be `$c : supportdyn<KeyedContainer<T#1, T#2>>`. To address this, we implement a rule that
`supportdyn<C<T#1>> <: C<supportdyn<T#1>>` where `C` has Container as an ancestor.

This is sound because `Container` is sealed s.t. only builtin collections are sub-classes, and for those, the only way they can be SDT (and hence have a non-empty intersection with dynamic) is if they contain only SDT elements (otherwise their getter could be invoked dynamically and yet return a non-SDT value).

Ideally, we'd like to apply this to all `Traversable`s, but is not sealed (via `ArrayIterator`), and so we are unsure that the rule is sound in that case.

This currently has a completeness issue where we raise a Container subtype to a Traversable super-type before supportdyn pushing. To be addressed in a subsequent commit.

Reviewed By: andrewjkennedy

Differential Revision: D43307827

fbshipit-source-id: 1c45806be8dd8eb515b8d0fc850bf1faf5e8c223