Like pushing incompleteness for expected types
Summary: When processing the "expected type", we apply like-pushing for Sound Dynamic. For example, when passing an argument to a function that expects `~vec<t>`, we don't just strip the `~` but push it in and return `vec<~t>` as the expected type, because `vec<~t> <: ~vec<t>`. But like-pushing is valid only if `t` is a dynamic-aware subtype of `dynamic`, so in the code we check for this with `is_supportdyn`. Unfortunately this isn't a complete test, for example `t` might be an unresolved type variable. In this case, rather than return an expected type `vec<t>` we remove the expected type entirely. But we also improve the heuristic by remembering if we've gone under a `supportdyn` type - in this case, there is no need to use the `is_supportdyn` test.

Differential Revision: D43319301

fbshipit-source-id: 863725a1269586f3156dc4f189d7b8f53d0303c5