Remove NoMockObjects flag from TypeConstraint
Summary:
Instead of puttinng a NoMockObjects flag in the type-constraint, set
an Attr on the class instead (the property of not having mock objects
belongs to the class, not the constraint). HHBBC can set the Attr
instead of setting the flag on TypeConstraint.

This flag is only really used in one place. We were checking in two
other places, but only if isThis() on the type-constraint was
true. The problem with this was that the flag is only set if the
type-constraint is AnnotType::Object, so those checks will always
fail. Remove them.

Reviewed By: jano

Differential Revision: D43372846

fbshipit-source-id: bce75a6282a6d488d61ae425f0a4692b8e6c82de