Fix hhvm/hackc common files to be able to build on mac
Summary:
Infer wants to be able to build and run hackc locally on Mac.  Most of hackc is in Rust so it builds fine but there are a few shared files so we need to leave this _APPLE_ #ifdef so it can build.

This isn't used by HHVM which no longer builds on Mac.

Reviewed By: ricklavoie

Differential Revision: D43370525

fbshipit-source-id: 6bad0218da4681f33c4c7458b7c047347b7c440e