Improve supportdyn pushing wrt Traversable
Summary: Add a flag to subtyping for classes to force like pushing, even when the subtype is not a Container. This is for the case where we have, e.g., `supportdyn<vec<t>> <: Traversable<supportdyn<t>>`. Here the inference algorithm first promotes the `vec` to `Traversable` and checks `supportdyn<Traversable<t>> <: Traversable<supportdyn<t>>` before trying to do supportdyn pushing to the `vec`. The flag indicates that we are in this situation and so it is ok to supportdyn push through the `Traversable`, which isn't generally sound, since user code can implement `Traversable`

Reviewed By: andrewjkennedy

Differential Revision: D43391212

fbshipit-source-id: 4bbe96358afd735a7f522a01e517785fbb18a8d4