Basic SDT analysis for top-level functions
Summary:
Adds constraint generation for what needs to be marked SDT if we start with the assumption that everything is already marked with SDT (we ignore the marking but take the consequences of the marking into account).

Currently, the constraints are only generated for functions.

The constraint is generated under two circumstances:
- If we call this function and if the function didn't support dynamic, i.e., the function type wasn't wrapped with `supportdyn<function ...>`, would we be able to typecheck the call? If not, the function has to be SDT.
- If the function is marked with `<<__DynamicallyCallable>>`, we cannot track the function calls statically, so we conclude that the function needs to be also marked with SDT.

Reviewed By: mheiber

Differential Revision: D43379776

fbshipit-source-id: 140843486ed4e569141f897a378384721c41fb4e