Fix QueryM isset mode
Summary:
The QueryM isset mode is used for expressions like `isset($a[5])`.  It was previously incorrectly ignored.  Fixed and added a test.

Also improved the test updated to match on words so that test commands don't have to include trailing tokens to make sure that they only match a unique string.

Reviewed By: davidpichardie

Differential Revision: D42443398

fbshipit-source-id: c67ab2dc7517841bf6e9a3ed091d677a785056a6