Thread the dynamic method check status to TASTs
Summary:
Saved environments stored inside TAST definitions is a section of the typing environment. This section didn't include whether a definition is checked under dynamic assumptions for support dynamic type.

This diff makes it so that this information is stored in the definition, so that it can be utilised during TAST processing.

Reviewed By: andrewjkennedy

Differential Revision: D43404453

fbshipit-source-id: 3fc3fc2b4267ac7af09fd09f52f5cdc52c364745