log the time for a server to finish retranslateall
Summary:
The BootTimer previously contained information on steps that happen during server initialization, and thus it included time for retranslateall in jumpstart mode, but not for the normal profiling-optimizing mode.
In this diff, we also log the time it takes to profile and retranslateall for non-jumpstarted servers. This is useful to estimate/track the capacity loss due to warmup.

Reviewed By: jano

Differential Revision: D43389605

fbshipit-source-id: 3b1a881597a48375b38b86ef0fca949fbd788b03