Fix codegen for inlined `transform.explicit` ctors and fields
Summary:
Our codegen inlines the `transform` of any ctor or struct field marked as `transform.explicit` with the intention that it behaves exactly as though we were making a function call.

Currently this isn't quite right; during a `transform` call the first thing we do is clone the incoming `impl pass` but in the inlined equivalent we operate directly on the mutable pass.

The consequence of this is that all subsequent transforms are using the mutated version of the pass. When the pass caries state this changes behaviour.

This diff fixes this by cloning the pass for both top-down and bottom-up passes, ensuring we don't leak modified state in subsequent transforms

Reviewed By: rjbailey

Differential Revision: D43393097

fbshipit-source-id: c6a9ec0bcaef2b5d44d279e7c9c4b3142158a565