Hover types under dynamic assumptions
Summary: Extend hover so that it shows the types of expressions under the "dynamic assumptions" check for SDT functions and methods. Currently this will only show if `--tast-under-dynamic` is enabled; also, if the types of static and dynamic passes match (or if no dynamic pass was required, because parameters are all enforced), then no additional information is displayed.

Reviewed By: mheiber, madgen

Differential Revision: D43237105

fbshipit-source-id: dcc146ac441597dfa14428621a211cd5160f7e71