Include saved state rollout flags in project_metadata
Summary:
This is gated under flag `project_metadata_w_flags`, which is set to false in JK for now.

Prior to this diff, the project metadata format is `<hhconfig_hash>-<hh_version_hash>`
With this diff it is `<hhconfig_hash>-<hh_version_hash>-<bit_array_of_rollout_flags>`.

This diff is verbose because we had to thread the rollout flags down to `GlobalOptions.t`.
Review `globalOptions.ml` first

Reviewed By: nt591

Differential Revision: D43352990

fbshipit-source-id: d0f1786209a47c87f6cdfae9e8e15a156c8e50be