Add logging for runtime-options on startup
Summary:
Uses the new Log() and HashAll() functions to add support for a new logging feature for runtime-options. Additionally uses the GetAllAsDynamic() function to replace startup logging of runtime options and admin handler reporting via `ini-get-all`.

Note that this behavior won't be identical to the existing `GetAll()` behavior as the new dynamic that is returned does not have a lot of the uninteresting PHP boilerplate that was encoded in the array returned by `GetAll()`.

Reviewed By: ricklavoie

Differential Revision: D42810671

fbshipit-source-id: 0d23053bee80f09dd6c39314f77f7f51944d60a8