Store native types in IniSetting
Summary: Most HDF (and INI) values are stored in C++ data structures, but the IniSetting class stores a map of callbacks which convert these types to their Hack equivalents. I've always found this interface clunky, it shouldn't be necessary to use VM types when generating a JSON encoding of the settings map or creating a hash.

Reviewed By: ricklavoie

Differential Revision: D42810673

fbshipit-source-id: 0231c1285058147f26c15a3b2a828f0ecf6835c5