Use Display format for Pos/RelativePath Debug impls
Summary: For a given span, we should only ever construct one representation of Pos (i.e., Large, Small, or Tiny), so it's not useful for the Debug representation to tell us which one we happened to use. Keep Debug output small for types containing positions or paths by using the briefer Display format.

Reviewed By: edwinsmith

Differential Revision: D43417276

fbshipit-source-id: 570b57e22d746da822d302eb9cf44a80f21362a3