Provide consts RelativePath::EMPTY and Pos::NONE
Summary:
Allow the empty path and position to be used in `const` contexts and constructed without allocating. The empty RelativePath should not allocate before this diff (because the empty PathBuf/OsString/Vec, AIUI, does not), but the empty Pos does, since it needs to construct an RcOc containing the empty RelativePath. If we make the RelativePath optional in the PosImpl::Tiny case, we can construct Pos::NONE in const contexts without allocating.

If we do that, then in order to implement `Pos::filename(&self) -> &RelativePath`, we must have some way of referencing a `'static` RelativePath (so that we can still return a value when the path in `self` is None). This constrains us to also make the PathBuf in RelativePath optional (since `PathBuf::new` is not a const fn).

Reviewed By: edwinsmith

Differential Revision: D43402966

fbshipit-source-id: 722a9b86acaccc951efae1ef4a20dcfc8f9a4009