Partially allow union simplification change for supportdyn
Summary:
{D43121847 (https://github.com/facebook/hhvm/commit/cd49bf2e547932e17aea74650e1e8de015b19e76)} removed `supportdyn` from the union simplification code for the case of `N<u> | N<v>` i.e. two instantiations of the same newtype, which allowed us to avoid the bad situation shown in the test case where an unawaited Awaitable object would be inserted into a keyset.

However, in a case with lots of nested array access and especially shapes, Hack's performance relies on the ability to push down unions into e.g. shared shape keys to prevent the types from growing unbounded.

`simplify_union` will return `None` in cases where it has decided not to simplify, and the newtype case unconditionally wraps the result of the `union` call (does not return `None` but rather a `ty`, see `union_lists`), which was the source of the problem with supportdyn addressed in D43121847 (https://github.com/facebook/hhvm/commit/cd49bf2e547932e17aea74650e1e8de015b19e76). This change peeks into the types and constructs a `supportdyn<union>` only in the cases where simplify_union would have returned a Some in the first place, and otherwise forwards along the refusal to simplify (None case).

Based on these observations, I'm not convinced union simplification for equal newtypes is actually correct in all cases. This would be a good test for enetsee's proposal for property-based testing.

Reviewed By: andrewjkennedy

Differential Revision: D43513084

fbshipit-source-id: 7472f98e9dda9869bd0daacf9edad6581ad70967