Use `ppx_transform`-derived traversal on expressions rather than hand-written one
Summary:
`naming_elab_const_expr` invalidates non-constant expressions appearing in constant contexts.

Currently the code performs a hand-written, top-down traversal (with wildcard matches!) of the expression during a bottom-up pass. This is confusing and hard to maintain - `ppx_transform` was introduced to eliminate exactly these kinds of traversals.

This diff refactors the pass, allowing us to use the pass handlers directly to acheive an equivalent outcome

Reviewed By: shayne-fletcher

Differential Revision: D43498181

fbshipit-source-id: d107058e47b09490fac6fb8c7aae5d6123832c1e