JK for ide_standalone
Summary:
Currently, clientLsp has a flag "serverless_ide = true/false" which it (1) uses to decide whether to launch clientIdeDaemon, (2) uses to decide whether to dispatch various requests to clientIdeDaemon or hh_server, (3) writes into logs.

In this diff I expand it into three states, not just a boolean. It's still used for all three purposes the same way. In future diffs it will also be used to decide whether further things (find-refs, find-all-squiggles) should be dispatched via RPC or shelling out.
```
type serverless_ide =
  | Serverless_not  (** all IDE requests are handled by RPC to hh_server *)
  | Serverless_with_rpc
      (** most IDE requests handled by clientIdeDaemon, some by RPC to hh_server *)
  | Serverless_with_shell
      (** most IDE requests handled by clientIdeDaemon, some by shelling-out to hh_server *)
```

Note: this changes the data-format of the scuba column `ide_serverless`. Previously it was `null` for non-LSP events and Serverless_not, and the number `1` for Serverless_with_rpc. This diff changes it to a normal, still `null` for non-LSP events, but now one of the three strings `Serverless_not, Serverless_with_rpc, Serverless_with_shell`. It's a bit yucky to change the format of a scuba column, but I think it'll be neater in the end.

I didn't bother adding a HackEventLogger.rollout_flag for it. That's because hh_server_events already has a column `serverless_ide` which is written to for all LSP events.

(Separately, for each individual LSP request, we send back a "powered-by" field to VSCode which says whether this particular request was handled by clientLsp or by clientIdeDaemon or by hh_server. This diff doesn't yet touch that).

To support this, I added a ServerLocalConfig flag. Previously it just had bool "ide_serverless". This diff adds an additional "ide_standalone" (backed by a JK which I already added) with the meaning:
* if `ide_standalone=true` -> `Serverless_with_shell`
* otherwise, if `ide_serverless=true` -> `Serverless_with_rpc`
* otherwise, `Serverless_not`

This flag can be set in the usual ways -- in hh.conf, with "--config ide_standalone=true", and with JK.

Reviewed By: nt591

Differential Revision: D43540114

fbshipit-source-id: 153367773ca6ed071b87bd79e2d6358c08ab5201