buck2 + Windows compatibility for thrift plugins
Summary:
In buck2, file paths can get very, very long. Specifically resolving includes can overflow the 260 MAX_PATH limit
because it concantenates the genrule pwd with the relative include path. We can work around this by absolutizing and normalizing
the include path, plus prefixing with `\\?\` for good measure, if a file we tested fails to be found the first time.

Note that these operations are very specific - neither absolute() nor lexically_normal() require the file to exist. Other transformations
like canonical() or weakly_canonical() require the file to exist and will throw if it does not. This way we can, in an exception-safe manner,
construct an alternate path to test (and exists() doesn't throw either).

Additionally, paths to executables cannot contain an `=` sign. If the cwd includes an `=` and buck attempts to construct a path to another binary via a relative path concatenated with the cwd, executing the binary will fail, even if that binary is at an absolute path without an `=` in it. As per precedent elsewhere, change `=` to `__` in the generated target names and fixup fallout from that.

Reviewed By: andrewjcg

Differential Revision: D43214613

fbshipit-source-id: 31b581099bec52f3e13736fe5a78fafd6be284ea