api for sdt_analysis: used by tast_check and hh_single_type_check
Summary:
# What

Give sdt_analysis a small API:
- One val for hh_single_type_check
- One val for our tast check

And introduce a noop tast check for sdt_analysis, which will gain interesting behavior later in the stack

# How

Use the default dune behavior of wrapping a library. Note: afaik there's no way to enforce such an API boundary with buck2.

# Why

shape_analysis exposed all its modules, which lead to hh_single_type_check and the logger reaching into ~5 different modules to pull out different bits and bobs. And there was duplicate (and accidentally different) logic w.r.t. the walker, the logger, and hh_single_type_check. I also found it hard to see what was API surface and what was implementation.

For sdt_analysis, I want to try having a small, clear API using the default dune behavior of wrapping.

# Why now

I will add more to this stack soon so that the logger persists constraints to disk.

I'm opening the diff now in order to put {D43444646} in context: the reason for implementing a tast_handler is so we can use it from a tast_check, and it minimizes discrepancies to have the *same* visitor in sdt_analysis regardless of whether invoked via `hh_single_type_check` or `hh`.

Reviewed By: madgen

Differential Revision: D43515394

fbshipit-source-id: 631fd52e1f73bb2f39a1f164644e35854b0110cd